adjusted exolore component to fix nav bar issue
