+++ client/src/App.js
@@ -1,17 +1,18 @@
 /** @format */
-
 import React from "react";
 import { MuiThemeProvider } from "@material-ui/core";
 import { BrowserRouter, Route, Redirect, Switch } from "react-router-dom";
 import Signup from "./pages/Signup";
+import Explore from './pages/explorer/explorer'
 import Signin from "./pages/Signin";
 import Flights from './pages/Flights/Flights'
 import Hotels from './pages/Hotels/Hotels'
 import Rent from './pages/Rent/Rent'
-
+import Profile from "./pages/Profile/Profile";
 import { theme } from "./themes/theme";
 
 import "./App.css";
+
 
 function App() {
 	return (
@@ -21,9 +22,11 @@ function App() {
 					<Redirect exact from="/" to="/signup" />
 					<Route path="/signup" component={Signup} />
 					<Route path="/signin" component={Signin} />
-					<Route path="/" component={Flights} exact />
+					<Route path="/explore" component={Explore} exact />
+					<Route path="/flights" component={Flights} exact />
 					<Route path="/hotel" component={Hotels} exact />
 					<Route path="/rent" component={Rent} exact />
+					<Route path="/profile" component={Profile} exact />
 				</Switch>
 			</BrowserRouter>
 		</MuiThemeProvider>
