+++ client/src/pages/Hotels/Hotels.js
@@ -1,8 +1,10 @@
 import React from 'react'
+import Header from '../../component/Header/Header'
 
 function Hotels() {
     return (
         <div>
+            <Header />
             <h1>This is the hotel page</h1>
         </div>
     )
