+++ client/src/pages/Profile/Profile.js
@@ -1,13 +1,14 @@
 import React from 'react'
 import Header from '../../component/Header/Header'
+import useStyles from "./style";
 
-function Rent() {
+function Profile() {
     return (
         <div>
             <Header />
-            <h1>This is rent page</h1>
+            <h1>Profile Page</h1>
         </div>
     )
 }
 
-export default Rent
+export default Profile
