+++ client/src/pages/Profile/style.js
@@ -1,0 +1,7 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+   
+  }));
+
+export default useStyles
