+++ client/src/pages/Signin.js
@@ -7,6 +7,6 @@ import SigninContainer from "./SigninContainer";
 const Signin = () => {
 	const [state, setState] = useState(false);
 	const handleRedirect = () => setState(true);
-	return state.redirect ? <Redirect to="/dashboard" /> : <SigninContainer redirect={handleRedirect} />
+	return state.redirect ? <Redirect to="/explore" /> : <SigninContainer redirect={handleRedirect} />
 };
 export default Signin;
